rename helper_functions.py to utils.py, fix a couple of formatting and a small numpy assignment deprecation
