+++ flexknot/likelihoods.py
@@ -4,10 +4,8 @@ Likelihoods using flex-knots.
 
 import numpy as np
 from scipy.special import erf, logsumexp
-from flexknot.helper_functions import (
-    get_x_nodes_from_theta,
-    get_theta_n,
-)
+from flexknot.utils import get_x_nodes_from_theta
+
 from flexknot.core import AdaptiveKnot, FlexKnot
